# Conflicts:
#	UJCSystem/ITJakub.Web.Hub/appsettings.LocalDebug.json
#	UJCSystem/ITJakub.Web.Hub/wwwroot/js/itjakub.ts
#	UJCSystem/ITJakub.Web.Hub/wwwroot/package.json
#	UJCSystem/ITJakub.Web.Hub/wwwroot/yarn.lock

+++ UJCSystem/Vokabular.MainService.Core/Works/Permission/DeleteRoleWork.cs
@@ -1,7 +1,7 @@
 ﻿using Ridics.Authentication.DataContracts;
 using Vokabular.DataEntities.Database.Repositories;
-using Vokabular.DataEntities.Database.UnitOfWork;
 using Vokabular.MainService.Core.Communication;
+using Vokabular.Shared.DataEntities.UnitOfWork;
 
 namespace Vokabular.MainService.Core.Works.Permission
 {
@@ -21,9 +21,12 @@ namespace Vokabular.MainService.Core.Works.Permission
         protected override void ExecuteWorkImplementation()
         {
             var group = m_permissionRepository.FindGroupByExternalId(m_roleId);
-            m_permissionRepository.Delete(group);
-            m_permissionRepository.Flush();
-
+            if (group != null)
+            {
+                m_permissionRepository.Delete(group);
+                m_permissionRepository.Flush();
+            }
+            
             var client = m_communicationProvider.GetAuthRoleApiClient();
             client.HttpClient.DeleteItemAsync<RoleContract>(m_roleId).GetAwaiter().GetResult();
         }
